fix: refactor upload modal
