+++ apps/web/components/Feature/Dashboard/CreatorProfile/index.tsx
@@ -1,13 +1,12 @@
 "use client";
 
-import React, { useMemo, useState, useCallback, useRef } from "react";
+import React, { useMemo, useState, useCallback } from "react";
 import { useTranslation } from "react-i18next";
 import {
   Container,
   Title,
   Card,
   Row,
-  Avatar,
   Fields,
   Action,
   Button,
@@ -16,22 +15,8 @@ import {
   InlineLabel,
   HeaderActions,
   NameBlock,
-  AvatarImage,
-  AvatarEditButton,
-  HiddenInput,
-  PhotoModalBody,
-  UploadDropZone,
-  UploadHint,
-  UploadOrText,
-  CropCanvas,
-  ModalActions,
-  ImagePreviewWrapper,
-  ImagePreview,
-  CropOverlay,
-  ZoomContainer,
-  ZoomSlider,
 } from "./styles";
-import { CREATOR_PROFILE } from "@/utils/translationKeys";
+import { AUTH, CREATOR_PROFILE } from "@/utils/translationKeys";
 import InputField from "@/components/UI/InputFields";
 import PasswordSection from "./PasswordSection";
 import CompanySection from "./CompanySection";
@@ -50,9 +35,9 @@ import { getProfileFields } from "@/utils/creatorProfilefields";
 import { PasswordState, ProfileForm } from "@/utils/creatorProfile";
 import { MODAL_ALIGN } from "@/utils/ui";
 import { GenericModal } from "@/components/UI/Modals";
-import { EditProfileIcon, SuccessArcIcon } from "@/assets/icons";
+import { SuccessArcIcon } from "@/assets/icons";
 import { useRouter } from "next/navigation";
-import { getCroppedImg } from "@/utils/image";
+import ImageUploader from "./ImageUploader";
 
 export default function CreatorProfile() {
   const { t } = useTranslation();
@@ -71,24 +56,6 @@ export default function CreatorProfile() {
   const [showDeleteModal, setShowDeleteModal] = useState(false);
   const [showDeleteSuccessModal, setShowDeleteSuccessModal] = useState(false);
   const [avatarImage, setAvatarImage] = useState<string | null>(null);
-  const [showPhotoModal, setShowPhotoModal] = useState(false);
-  const [photoStep, setPhotoStep] = useState<"upload" | "edit">("upload");
-  const [pendingImage, setPendingImage] = useState<string | null>(null);
-
-  const [zoom, setZoom] = useState(1);
-  const [position, setPosition] = useState({ x: 0, y: 0 });
-  const [isDraggingImage, setIsDraggingImage] = useState(false);
-
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const previewFrameRef = useRef<HTMLDivElement | null>(null);
-  const cropAreaRef = useRef<HTMLDivElement | null>(null);
-
-  const dragStateRef = useRef<{
-    startX: number;
-    startY: number;
-    originX: number;
-    originY: number;
-  } | null>(null);
 
   const dirty = useMemo(() => {
     const formChanged = JSON.stringify(form) !== JSON.stringify(saved);
@@ -155,110 +122,8 @@ export default function CreatorProfile() {
 
   const handleDeleteSuccessClose = () => {
     setShowDeleteSuccessModal(false);
-    router.push("/auth/login");
+    router.push(AUTH.login);
   };
-
-  const openPhotoModal = useCallback(() => {
-    setShowPhotoModal(true);
-    if (avatarImage) {
-      setPendingImage(avatarImage);
-      setPhotoStep("edit");
-    } else {
-      setPhotoStep("upload");
-    }
-  }, [avatarImage]);
-
-  const closePhotoModal = useCallback(() => {
-    setShowPhotoModal(false);
-    setPhotoStep("upload");
-    setPendingImage(null);
-    setZoom(1);
-    setPosition({ x: 0, y: 0 });
-    setIsDraggingImage(false);
-    dragStateRef.current = null;
-  }, []);
-
-  const handleSelectFile = useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => {
-      const file = event.target.files?.[0];
-      if (!file) return;
-
-      event.target.value = "";
-
-      const reader = new FileReader();
-      reader.onload = () => {
-        const result = reader.result;
-        if (typeof result !== "string") return;
-
-        setPendingImage(result);
-        setPhotoStep("edit");
-        setZoom(1);
-        setPosition({ x: 0, y: 0 });
-      };
-
-      reader.readAsDataURL(file);
-    },
-    [],
-  );
-
-  const handleMouseDown = (e: React.MouseEvent) => {
-    if (!pendingImage) return;
-
-    dragStateRef.current = {
-      startX: e.clientX,
-      startY: e.clientY,
-      originX: position.x,
-      originY: position.y,
-    };
-
-    setIsDraggingImage(true);
-  };
-
-  const handleMouseMove = (e: React.MouseEvent) => {
-    if (!dragStateRef.current) return;
-
-    const dx = e.clientX - dragStateRef.current.startX;
-    const dy = e.clientY - dragStateRef.current.startY;
-
-    setPosition({
-      x: dragStateRef.current.originX + dx,
-      y: dragStateRef.current.originY + dy,
-    });
-  };
-
-  const stopDragging = () => {
-    dragStateRef.current = null;
-    setIsDraggingImage(false);
-  };
-
-  const applyCrop = useCallback(async () => {
-    if (!pendingImage) return;
-
-    const frame = previewFrameRef.current;
-    if (!frame) return;
-
-    const { width, height } = frame.getBoundingClientRect();
-
-    const croppedImage = await getCroppedImg(
-      pendingImage,
-      width,
-      height,
-      {
-        x: 0,
-        y: 0,
-        width,
-        height,
-      },
-      {
-        x: position.x / zoom,
-        y: position.y / zoom,
-      },
-      zoom,
-    );
-
-    setAvatarImage(croppedImage);
-    closePhotoModal();
-  }, [pendingImage, position, zoom, closePhotoModal]);
 
   return (
     <Container>
@@ -278,37 +143,15 @@ export default function CreatorProfile() {
 
       <Card>
         <Row>
-          <Avatar
-            role="button"
-            aria-label={t("creatorProfile.editPhoto")}
-            onClick={openPhotoModal}
-            tabIndex={0}
-            onKeyDown={(event) => {
-              if (event.key === "Enter" || event.key === " ") {
-                event.preventDefault();
-                openPhotoModal();
-              }
-            }}
-          >
-            {avatarImage ? (
-              <AvatarImage
-                src={avatarImage}
-                alt={t("creatorProfile.profilePhotoAlt")}
-              />
-            ) : (
-              <MonoText $use="Heading2">{getInitial(email)}</MonoText>
-            )}
-            <AvatarEditButton
-              type="button"
-              aria-label={t("creatorProfile.editPhoto")}
-              onClick={(event) => {
-                event.stopPropagation();
-                openPhotoModal();
-              }}
-            >
-              <EditProfileIcon width={36} height={36} />
-            </AvatarEditButton>
-          </Avatar>
+          <ImageUploader
+            image={avatarImage}
+            fallback={getInitial(email)}
+            alt={t("creatorProfile.profilePhotoAlt")}
+            uploadTitle={t("creatorProfile.uploadPhotoTitle")}
+            editTitle={t("creatorProfile.editPhotoTitle")}
+            onChange={setAvatarImage}
+          />
+
           <NameBlock>
             <MonoText $use="Heading3">{name}</MonoText>
             <MonoText $use="Body_Medium" color={COLORS.neutral.GRAY}>
@@ -343,86 +186,7 @@ export default function CreatorProfile() {
       <CompanySection form={form} onChange={onChange} t={t} />
       <PaymentSection form={form} onChange={onChange} t={t} />
       <DeleteSection onDelete={() => setShowDeleteModal(true)} />
-      <GenericModal
-        visible={showPhotoModal}
-        title={
-          photoStep === "upload"
-            ? t("creatorProfile.uploadPhotoTitle")
-            : t("creatorProfile.editPhotoTitle")
-        }
-        onClose={closePhotoModal}
-        textAlign={MODAL_ALIGN.START}
-        width="630px"
-      >
-        <PhotoModalBody>
-          <HiddenInput
-            ref={fileInputRef}
-            type="file"
-            accept="image/*"
-            onChange={handleSelectFile}
-          />
-          {photoStep === "upload" ? (
-            <UploadDropZone>
-              <UploadHint>{t("creatorProfile.dragPhotoHere")}</UploadHint>
-              <UploadOrText>{t("creatorProfile.or")}</UploadOrText>
-              <GenericButton
-                variant={VARIANT.PRIMARY}
-                onClick={() => fileInputRef.current?.click()}
-              >
-                {t("creatorProfile.choosePhoto")}
-              </GenericButton>
-            </UploadDropZone>
-          ) : (
-            <>
-              <CropCanvas>
-                <ImagePreviewWrapper
-                  ref={previewFrameRef}
-                  onMouseDown={handleMouseDown}
-                  onMouseMove={handleMouseMove}
-                  onMouseUp={stopDragging}
-                  onMouseLeave={stopDragging}
-                >
-                  {pendingImage && (
-                    <ImagePreview
-                      src={pendingImage}
-                      alt={t("creatorProfile.profilePhotoAlt")}
-                      $x={position.x}
-                      $y={position.y}
-                      $zoom={zoom}
-                      $isDragging={isDraggingImage}
-                    />
-                  )}
 
-                  <CropOverlay ref={cropAreaRef} />
-                </ImagePreviewWrapper>
-              </CropCanvas>
-
-              <ZoomContainer>
-                <ZoomSlider
-                  type="range"
-                  min={1}
-                  max={3}
-                  step={0.01}
-                  value={zoom}
-                  onChange={(e) => setZoom(Number(e.target.value))}
-                />
-              </ZoomContainer>
-
-              <ModalActions>
-                <GenericButton
-                  variant={VARIANT.SECONDARY}
-                  onClick={closePhotoModal}
-                >
-                  {t("common.cancel")}
-                </GenericButton>
-                <GenericButton variant={VARIANT.PRIMARY} onClick={applyCrop}>
-                  {t("creatorProfile.apply")}
-                </GenericButton>
-              </ModalActions>
-            </>
-          )}
-        </PhotoModalBody>
-      </GenericModal>
       <GenericModal
         visible={showPassword}
         title={t(CREATOR_PROFILE.changePassword)}
@@ -457,7 +221,7 @@ export default function CreatorProfile() {
         message={t("creatorProfile.passwordSuccessMessage")}
         confirmLabel={t("nav.login")}
         onClose={() => setShowPasswordSuccessModal(false)}
-        onConfirm={() => router.push("/auth/login")}
+        onConfirm={() => router.push(AUTH.login)}
         width="480px"
         showCloseButton={false}
       />
